requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
Flask-Login==0.6.3
Flask-Migrate==4.0.5
python-dotenv==1.0.0
werkzeug==3.0.1
